requirements.txt
dacite
matplotlib
numpy
omegaconf
opencv_python==4.6.0.66
pandas
plotly
imageio
PyYAML
scipy
tqdm
pydantic
rich
hydra-core
--index-url https://download.pytorch.org/whl/cu118
torch
torchaudio
torchvision
einops
tensorboard
torchmetrics==0.11.4
